Refactor the sinks to use the new interface
