+++ vpt/sinks/sqlite_store/__init__.py
@@ -18,6 +18,8 @@ class SQLiteStore(SinkBase):
     def __init__(self, db_path: str, mouse_source: SourceBase,
                  keyboard_source: SourceBase, engagement_source: SourceBase):
         '''Create a database or open an existing one.'''
+        self.stopped = True
+        self.sources = [mouse_source, keyboard_source, engagement_source]
         self.connection = sqlite3.connect(db_path)
         cur = self.connection.cursor()
         cur.execute('''
@@ -47,9 +49,9 @@ class SQLiteStore(SinkBase):
         self.connection.commit()
         cur.close()
 
-        mouse_source.get_data_stream().subscribe(self.store_mouse_event)
-        keyboard_source.get_data_stream().subscribe(self.store_key_event)
-        engagement_source.get_data_stream().subscribe(self.store_engagement)
+        mouse_source.output.subscribe(self.store_mouse_event)
+        keyboard_source.output.subscribe(self.store_key_event)
+        engagement_source.output.subscribe(self.store_engagement)
 
     def __del__(self):
         '''Clean up resources.'''
